refactor: move readiness implementation to @percy/sdk-utils (PER-7348)
The 556-line readiness implementation (MutationObserver, PerformanceObservers,
font/image checks, presets, abort handling) lived in @percy/dom because the
code physically runs in the browser via the @percy/dom bundle. But the
SDK-facing concerns — config precedence (getReadinessConfig), in-browser
invoker script emission (waitForReadyScript), kill-switch (isReadinessDisabled)
— already lived in @percy/sdk-utils since CLI #2184. Splitting "the readiness
implementation" from "the readiness orchestration" across two packages is
confusing for maintainers.

This commit consolidates the source of truth in @percy/sdk-utils:

- Moved packages/dom/src/readiness.js → packages/sdk-utils/src/readiness-browser.js
- Moved packages/dom/test/readiness-helpers.test.js → packages/sdk-utils/test/
  (the helpers tested are internal to the moved file)
- packages/dom/src/index.js now re-exports waitForReady from sdk-utils
- packages/sdk-utils/package.json adds the file to `files` and adds an
  exports map entry: `./readiness-browser`

Why the file lives as a source file in sdk-utils but is consumed via
@percy/dom: the code uses browser globals (MutationObserver, document,
performance, etc.) so it cannot run in Node. @percy/dom's rollup build
inlines it into the browser bundle that ships via fetchPercyDOM(). End
users of @percy/dom get a self-contained bundle — no runtime dep on
sdk-utils — because the bundle is already inlined. SDKs that already
depend on sdk-utils for the Node-side helpers can now also access the
browser-side source directly via `@percy/sdk-utils/readiness-browser`
if they ever need to inject it themselves.

Behavior verified: @percy/dom's rebuilt dist/bundle.js still contains
`function waitForReady` and the readiness implementation inlined; the
public `PercyDOM.waitForReady` global remains identical for every SDK
caller (no SDK changes required).
