+++ packages/dom/src/index.js
@@ -8,4 +8,9 @@ export {
 
 export { loadAllSrcsetLinks } from './serialize-image-srcset';
 
-export { waitForReady } from './readiness';
+// Source of truth lives in @percy/sdk-utils. @percy/dom re-exports it here
+// so the browser bundle continues to attach `PercyDOM.waitForReady` as a
+// global — SDK callers (cypress, ember, puppeteer, etc.) keep working
+// unchanged. Moved here to consolidate readiness ownership in the
+// SDK-facing package (PER-7348).
+export { waitForReady } from '@percy/sdk-utils/readiness-browser';
